improve KCL with new technic slots. made some renaming add a lot of usefull function in KCLDetect KCLButton and SlideButton will be remove in next commit

+++ kcl/kclinputevent.cpp
@@ -1,6 +1,6 @@
 #include "kclinputevent.h"
 
-KCLInputEvent::KCLInputEvent( unsigned short type_,unsigned long tvSec_, unsigned long tvUsec_, unsigned short code_, unsigned short value_)
+KCLInputEvent::KCLInputEvent( int type_,unsigned long tvSec_, unsigned long tvUsec_, unsigned short code_, unsigned short value_)
     :QEvent(QEvent::Type(QEvent::User+type_))
 {
     struct timeval time;
